Use double for universal time
curtime is reset on new maps while our universal time doesn't (so cross-map timers can run), and once it reaches 262144 seconds increments become greater than the tick interval so timers break

+++ src/cs2fixes.cpp
@@ -60,7 +60,7 @@
 
 #include "tier0/memdbgon.h"
 
-float g_flUniversalTime;
+double g_flUniversalTime;
 float g_flLastTickedTime;
 bool g_bHasTicked;
