+++ src/zombiereborn.h
@@ -201,7 +201,7 @@ public:
 	static void RemoveAllTimers();
 
 private:
-	static float s_flNextExecution;
+	static double s_flNextExecution;
 	static CZRRegenTimer *s_vecRegenTimers[MAXPLAYERS];
 	int m_iRegenAmount;
 	CHandle<CCSPlayerPawn> m_hPawnHandle;
